requirements.txt
Django==5.0.3
django-cors-headers==4.4.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
